Move OsVersions to os package
